+++ userservices/service/context.go
@@ -1,10 +1,10 @@
 package service
 
 import (
+	confirmationStore "github.com/tidepool-org/platform/confirmation/store"
 	dataservicesClient "github.com/tidepool-org/platform/dataservices/client"
 	messageStore "github.com/tidepool-org/platform/message/store"
 	metricservicesClient "github.com/tidepool-org/platform/metricservices/client"
-	notificationStore "github.com/tidepool-org/platform/notification/store"
 	permissionStore "github.com/tidepool-org/platform/permission/store"
 	profileStore "github.com/tidepool-org/platform/profile/store"
 	"github.com/tidepool-org/platform/service"
@@ -20,8 +20,8 @@ type Context interface {
 	UserServicesClient() userservicesClient.Client
 	DataServicesClient() dataservicesClient.Client
 
+	ConfirmationStoreSession() confirmationStore.Session
 	MessageStoreSession() messageStore.Session
-	NotificationStoreSession() notificationStore.Session
 	PermissionStoreSession() permissionStore.Session
 	ProfileStoreSession() profileStore.Session
 	SessionStoreSession() sessionStore.Session