+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## HEAD
 
+- Move current notification functionality to confirmation package
 - Refactor version package to align with new application package
 - Refactor common functionality in service package into application package
 - Update service package to reflect new application package